+++ src/ClassFactory.php
@@ -69,7 +69,7 @@ abstract class ClassFactory
          * Evaluate states that are closures and group all resulting states per property
          *
          *  $this->states = [
-         *      ['id' => 1, 'key' => 'a', 'prop' => OtherClassFactory::new()],
+         *      ['id' => 1, 'key' => 'a', 'prop' => OtherClassFactory::new(), 'action' = ClosureValue::of(fn () => 123)],
          *      fn () => ['id' => 2],
          *      ['key' => fn ($attrs) => $attrs['key'] . 'b' . $attrs['id']],
          *      ['prop' => fn ($attrs) => $attrs['prop']->someState()],
@@ -79,6 +79,7 @@ abstract class ClassFactory
          *      'id' => [1, 2],
          *      'key' => ['a', fn ($attrs) => $attrs['key'] . 'b' . $attrs['id']],
          *      'prop' => [OtherClassFactory::new(), fn ($attrs) => $attrs['prop']->someState()],
+         *      'action' => [ClosureValue::of(fn () => 123)],
          *  ];
          */
         $collapsedState = array_reduce(
@@ -105,12 +106,14 @@ abstract class ClassFactory
          *      'id' => [1, 2],
          *      'key' => ['a', fn ($attrs) => $attrs['key'] . 'b' . $attrs['id']],
          *      'prop' => [OtherClassFactory::new(), fn ($attrs) => $attrs['prop']->someState()],
+         *      'action' => [ClosureValue::of(fn () => 123)],
          *  ];
          * turns into
          *  $collapsedState = [
          *      'id' => 2,
          *      'key' => ['a', fn ($attrs) => $attrs['key'] . 'b' . $attrs['id']],
          *      'prop' => [OtherClassFactory::new(), fn ($attrs) => $attrs['prop']->someState()],
+         *      'action' => ClosureValue::of(fn () => 123),
          *  ];
          *
          * This way we collapse everything we can before evaluating closures, so they get the most up-to-date state
@@ -128,12 +131,14 @@ abstract class ClassFactory
          *      'id' => 2,
          *      'key' => ['a', fn ($attrs) => $attrs['key'] . 'b' . $attrs['id']],
          *      'prop' => [OtherClassFactory::new(), fn ($attrs) => $attrs['prop']->someState()],
+         *      'action' => ClosureValue::of(fn () => 123),
          *  ];
          * turns into
          *  $collapsedState = [
          *      'id' => 2,
          *      'key' => 'ab2',
          *      'prop' => OtherClassFactory::new()->someState(),
+         *      'action' => ClosureValue::of(fn () => 123),
          *  ];
          */
         foreach ($collapsedState as $key => $values) {
@@ -152,15 +157,39 @@ abstract class ClassFactory
          *      'id' => 2,
          *      'key' => 'ab2',
          *      'prop' => OtherClassFactory::new()->someState(),
+         *      'action' => ClosureValue::of(fn () => 123),
          *  ];
          * turns into
          *  $collapsedState = [
          *      'id' => 2,
          *      'key' => 'ab2',
          *      'prop' => OtherClassFactory::new()->someState()->make(),
+         *      'action' => ClosureValue::of(fn () => 123),
          *  ];
          */
         $collapsedState = array_map($this->makeIfFactory(...), $collapsedState);
+
+        /**
+         * Unwrap closure values.
+         *
+         *  $collapsedState = [
+         *      'id' => 2,
+         *      'key' => 'ab2',
+         *      'prop' => OtherClassFactory::new()->someState()->make(),
+         *      'action' => ClosureValue::of(fn () => 123),
+         *  ];
+         * turns into
+         *  $collapsedState = [
+         *      'id' => 2,
+         *      'key' => 'ab2',
+         *      'prop' => OtherClassFactory::new()->someState()->make(),
+         *      'action' => fn () => 123,
+         *  ];
+         */
+        $collapsedState = array_map(
+            fn ($value) => $value instanceof ClosureValue ? $value->value : $value,
+            $collapsedState,
+        );
 
         return array_intersect_key($collapsedState, $definedProperties);
     }