+++ tests/TestClasses/NamedAction.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace EKvedaras\ClassFactory\Tests\TestClasses;
+
+use Closure;
+
+class NamedAction
+{
+    public function __construct(
+        public readonly string $name,
+        public readonly Closure $action,
+    ) {
+    }
+}